backend/ML_Model/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
Pillow